Build real Django 3.2 project + 4 quality tools for PYTHON_3.8
Same philosophy as other branches. Targets Django 3.2 LTS. Differentiator
versus the 3.7 branch: calculate_order_total's quantity/unit_price_cents
are positional-only parameters (PEP 570, /), the discount is computed
with a walrus-operator expression (PEP 572, :=), and validation error
messages use the f-string = debug specifier - all new exactly in
Python 3.8.

quality/ has 4 tools, matching Golden_Repo_Lite's Python_3.8 folder
exactly (drops beniget/jscpd, adds cognitive-ast/crosshair/radon-lizard
versus 3.7's set): cognitive-ast, crosshair, pydriller (a real
`python manage.py mine_history` Django command), and radon-lizard - all
against the real catalog/pricing.py.

Verified all .py files compile cleanly under Python 3.11, though no
actual Python 3.8 interpreter is available on this machine.

+++ catalog/migrations/0001_initial.py
@@ -9,21 +9,13 @@ class Migration(migrations.Migration):
 
     operations = [
         migrations.CreateModel(
-            name="Product",
+            name='Product',
             fields=[
-                (
-                    "id",
-                    models.BigAutoField(
-                        auto_created=True,
-                        primary_key=True,
-                        serialize=False,
-                        verbose_name="ID",
-                    ),
-                ),
-                ("name", models.CharField(max_length=200)),
-                ("price_cents", models.PositiveIntegerField()),
-                ("in_stock", models.BooleanField(default=True)),
-                ("created_at", models.DateTimeField(auto_now_add=True)),
+                ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('name', models.CharField(max_length=200)),
+                ('price_cents', models.PositiveIntegerField()),
+                ('in_stock', models.BooleanField(default=True)),
+                ('created_at', models.DateTimeField(auto_now_add=True)),
             ],
         ),
     ]
